FIX MINOR ERRORS
